Revert "SERVER-69767 Robustify feature flag checks for columnstore index tests"
This reverts commit 31526b6b7da43fff942a8fb184dab2e9b8f8b677.

+++ jstests/core/columnstore_eligibility.js
@@ -6,16 +6,23 @@
  *   does_not_support_stepdowns,
  *   # Cannot run aggregate with explain in a transaction.
  *   does_not_support_transactions,
+ *   # columnstore indexes are new in 6.2.
  *   requires_fcv_62,
  *   uses_column_store_index,
- *   featureFlagColumnstoreIndexes,
- *   featureFlagSbeFull,
  * ]
  */
 (function() {
 "use strict";
 
 load("jstests/libs/analyze_plan.js");
+load("jstests/libs/sbe_util.js");  // For checkSBEEnabled.
+
+const columnstoreEnabled =
+    checkSBEEnabled(db, ["featureFlagColumnstoreIndexes", "featureFlagSbeFull"]);
+if (!columnstoreEnabled) {
+    jsTestLog("Skipping columnstore index validation test since the feature flag is not enabled.");
+    return;
+}
 
 const coll = db.columnstore_eligibility;
 coll.drop();
