+++ jstests/core/columnstore_large_array_index_correctness.js
@@ -3,15 +3,23 @@
  * @tags: [
  *   # columnstore indexes are new in 6.1.
  *   requires_fcv_61,
- *   uses_column_store_index,
- *   featureFlagColumnstoreIndexes,
- *   featureFlagSbeFull,
+ *   # Columnstore indexes are incompatible with clustered collections.
+ *   incompatible_with_clustered_collection,
  * ]
  */
 (function() {
 "use strict";
 
 load("jstests/libs/analyze_plan.js");  // For "planHasStage."
+load("jstests/libs/sbe_util.js");      // For "checkSBEEnabled.""
+
+const columnstoreEnabled = checkSBEEnabled(
+    db, ["featureFlagColumnstoreIndexes", "featureFlagSbeFull"], true /* checkAllNodes */);
+if (!columnstoreEnabled) {
+    jsTestLog(
+        "Skipping columnstore large array index validation test since the feature flag is not enabled.");
+    return;
+}
 
 const coll = db.columnstore_index_correctness;
 coll.drop();
